store QueryDataSetSources & QueryNamespaceValues as string arrays
QueryDataSetSources & QueryNamespaceValues contain a list of values
that at the moment are stored, into rdf db, as a new measurement for
each value from the list. This causes problems when trying to make a
query that return all of these values as a single row for each
Query Context. 
To fix this, we will store store just one measurement
for QueryDataSetSources and one for QueryNamespaceValues by storing the
values as comma separated values.
example: [ns1, ns2, ns3]